# openapi.yml
openapi: "3.0.3"
info:
  title: コンビニトイレマップ API
  description: |
    コンビニのトイレ位置情報を共有・検索するWebアプリケーションのAPI。
    認証はフロントから Cognito SDK を直接利用。
    バックエンドは API Gateway の Cognito Authorizer で JWT を検証。
  version: "1.0.0"

servers:
  - url: https://api.example.com/v1
    description: Production (API Gateway + Lambda)

tags:
  - name: toilets
    description: トイレ情報の登録・取得
  - name: votes
    description: トイレ情報への投票
  - name: images
    description: 画像アップロード
  - name: notifications
    description: プッシュ通知・デバイス登録
  - name: location
    description: 位置情報チェック
  - name: navigation
    description: ナビゲーション

paths:
  # ============================================================
  # 画像アップロード（S3 presigned URL 発行）
  # ============================================================
  /images/presigned-url:
    post:
      tags: [images]
      summary: S3 presigned URL を発行
      operationId: createPresignedUrl
      security:
        - cognitoAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresignedUrlRequest"
      responses:
        "200":
          description: presigned URL 発行成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignedUrlResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ============================================================
  # トイレ情報
  # ============================================================
  /toilets:
    post:
      tags: [toilets]
      summary: トイレ情報を投稿
      description: |
        コンビニの店舗情報とトイレ情報をまとめて投稿する。
        画像は事前に S3 へアップロードし、取得した imageKey を指定する。
      operationId: createToilet
      security:
        - cognitoAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToiletRequest"
      responses:
        "201":
          description: 投稿成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Toilet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [toilets]
      summary: 指定範囲内のトイレ一覧を取得（マップ表示用）
      operationId: listToilets
      security:
        - cognitoAuth: []
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
          description: 中心緯度
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
          description: 中心経度
        - name: radius
          in: query
          required: false
          schema:
            type: integer
            default: 1000
            minimum: 100
            maximum: 5000
          description: 検索半径（メートル）
      responses:
        "200":
          description: トイレ一覧
          content:
            application/json:
              schema:
                type: object
                required: [toilets]
                properties:
                  toilets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Toilet"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ============================================================
  # トイレ虚偽報告
  # ============================================================
  /toilets/{toiletId}/reports:
    post:
      tags: [toilets]
      summary: トイレの虚偽報告を送信
      description: |
        トイレが存在しない等の虚偽報告を送信する。
        同一ユーザーの重複報告は 409 Conflict を返す。
        報告が3件に達したらトイレは自動削除される。
      operationId: createReport
      security:
        - cognitoAuth: []
      parameters:
        - $ref: "#/components/parameters/toiletId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportRequest"
      responses:
        "201":
          description: 報告成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateReportResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 重複報告
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /toilets/{toiletId}/reports/count:
    get:
      tags: [toilets]
      summary: トイレの報告件数を取得
      operationId: getReportCount
      security:
        - cognitoAuth: []
      parameters:
        - $ref: "#/components/parameters/toiletId"
      responses:
        "200":
          description: 報告件数
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportCountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /toilets/{toiletId}:
    delete:
      tags: [toilets]
      summary: トイレ情報を削除
      operationId: deleteToilet
      security:
        - cognitoAuth: []
      parameters:
        - $ref: "#/components/parameters/toiletId"
      responses:
        "204":
          description: 削除成功
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [toilets]
      summary: トイレ詳細を取得
      operationId: getToilet
      security:
        - cognitoAuth: []
      parameters:
        - $ref: "#/components/parameters/toiletId"
      responses:
        "200":
          description: トイレ詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Toilet"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ============================================================
  # 投票
  # ============================================================
  /toilets/{toiletId}/votes:
    post:
      tags: [votes]
      summary: トイレ情報に投票
      operationId: createVote
      security:
        - cognitoAuth: []
      parameters:
        - $ref: "#/components/parameters/toiletId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVoteRequest"
      responses:
        "201":
          description: 投票成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vote"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [votes]
      summary: トイレの投票一覧を取得
      operationId: listVotes
      security:
        - cognitoAuth: []
      parameters:
        - $ref: "#/components/parameters/toiletId"
      responses:
        "200":
          description: 投票一覧
          content:
            application/json:
              schema:
                type: object
                required: [votes]
                properties:
                  votes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vote"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ============================================================
  # プッシュ通知サブスクリプション
  # ============================================================
  /subscriptions:
    post:
      tags: [notifications]
      summary: デバイストークンを登録
      description: |
        プッシュ通知を受け取るためのデバイストークンを登録する。
        iOS / Android / Web (Web Push) に対応。
      operationId: createSubscription
      security:
        - cognitoAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
      responses:
        "201":
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /subscriptions/{subscriptionId}:
    delete:
      tags: [notifications]
      summary: デバイストークンを削除
      operationId: deleteSubscription
      security:
        - cognitoAuth: []
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: サブスクリプションID
      responses:
        "204":
          description: 削除成功
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ============================================================
  # ナビゲーション
  # ============================================================
  /navigation/route:
    post:
      tags: [navigation]
      summary: 徒歩ルートを計算
      description: |
        現在位置から目的地までの徒歩経路を計算し、GeoJSON LineString と AI 生成の案内文を返す。
      operationId: calculateNavigationRoute
      security:
        - cognitoAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NavigationRouteRequest"
      responses:
        "200":
          description: ルート計算成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NavigationRouteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ============================================================
  # 位置情報チェック（プッシュ通知トリガー）
  # ============================================================
  /location/check:
    post:
      tags: [location]
      summary: 現在位置を送信し周辺トイレをチェック
      description: |
        クライアントから定期的に現在位置を送信する。
        周辺 1km 以内にトイレが 0 件の場合、プッシュ通知で警告を送信し
        レスポンスでも通知済みフラグを返す。
      operationId: checkLocation
      security:
        - cognitoAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationCheckRequest"
      responses:
        "200":
          description: チェック結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationCheckResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

components:
  # ============================================================
  # セキュリティスキーム
  # ============================================================
  securitySchemes:
    cognitoAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Amazon Cognito が発行する ID Token / Access Token

  # ============================================================
  # パラメータ
  # ============================================================
  parameters:
    toiletId:
      name: toiletId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: トイレID

  # ============================================================
  # 共通レスポンス
  # ============================================================
  responses:
    BadRequest:
      description: リクエスト不正
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: リソースの競合
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ============================================================
  # スキーマ
  # ============================================================
  schemas:
    # --- ナビゲーション ---
    Coordinate:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
          description: 緯度
        lng:
          type: number
          format: double
          description: 経度

    NavigationRouteRequest:
      type: object
      required: [origin, destination, dest_name]
      properties:
        origin:
          $ref: "#/components/schemas/Coordinate"
        destination:
          $ref: "#/components/schemas/Coordinate"
        dest_name:
          type: string
          description: "目的地の名称（提案文生成に使用）"

    GeoJSONLineString:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [LineString]
        coordinates:
          type: array
          items:
            type: array
            items:
              type: number
              format: double

    NavigationRouteResponse:
      type: object
      required: [distance_meters, duration_seconds, polyline, suggestion_text]
      properties:
        distance_meters:
          type: number
          format: double
          description: 距離（メートル）
        duration_seconds:
          type: number
          format: double
          description: 所要時間（秒）
        polyline:
          $ref: "#/components/schemas/GeoJSONLineString"
        suggestion_text:
          type: string
          description: AI 生成の案内文

    # --- 共通 ---
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "リクエストが不正です"

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: "処理が完了しました"

    # --- 報告 ---
    CreateReportRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          enum:
            - not_exists
            - wrong_info
            - other
          description: "報告理由"
        comment:
          type: string
          maxLength: 500
          description: "コメント（任意）"

    CreateReportResponse:
      type: object
      required: [message, reportCount]
      properties:
        message:
          type: string
          example: "報告を受け付けました"
        reportCount:
          type: integer
          description: 現在の報告件数

    ReportCountResponse:
      type: object
      required: [count]
      properties:
        count:
          type: integer
          description: 報告件数

    # --- プッシュ通知 ---
    CreateSubscriptionRequest:
      type: object
      required: [platform, token]
      properties:
        platform:
          type: string
          enum:
            - ios
            - android
            - web
          description: "プラットフォーム種別"
        token:
          type: string
          description: |
            デバイストークン。
            Web Push の場合は PushSubscription を JSON 文字列化したもの。

    Subscription:
      type: object
      required: [subscriptionId, platform, createdAt]
      properties:
        subscriptionId:
          type: string
          format: uuid
        platform:
          type: string
          enum: [ios, android, web]
        createdAt:
          type: string
          format: date-time

    LocationCheckRequest:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
          description: 現在の緯度
        lng:
          type: number
          format: double
          description: 現在の経度

    LocationCheckResponse:
      type: object
      required: [nearbyCount, notified]
      properties:
        nearbyCount:
          type: integer
          description: 周辺 1km 以内のトイレ件数
        notified:
          type: boolean
          description: プッシュ通知を送信したか
        message:
          type: string
          description: 通知メッセージ（通知した場合のみ）

    # --- 投票 ---
    CreateVoteRequest:
      type: object
      required:
        - toiletType
        - requiresPermission
      properties:
        toiletType:
          type: string
          enum: [shared, separated]
        requiresPermission:
          type: boolean
        note:
          type: string
          maxLength: 500
        imageKey:
          type: string

    Vote:
      type: object
      required:
        - toiletId
        - userId
        - toiletType
        - requiresPermission
        - createdAt
      properties:
        toiletId:
          type: string
          format: uuid
        userId:
          type: string
        toiletType:
          type: string
          enum: [shared, separated]
        requiresPermission:
          type: boolean
        note:
          type: string
        imageUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # --- 画像 ---
    PresignedUrlRequest:
      type: object
      required: [contentType]
      properties:
        contentType:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/webp
          description: アップロードする画像のMIMEタイプ

    PresignedUrlResponse:
      type: object
      required: [uploadUrl, imageKey, expiresIn]
      properties:
        uploadUrl:
          type: string
          format: uri
          description: S3 presigned PUT URL
        imageKey:
          type: string
          description: S3オブジェクトキー（投稿時に指定）
        expiresIn:
          type: integer
          description: URL有効期限（秒）
          example: 300

    # --- トイレ ---
    CreateToiletRequest:
      type: object
      required:
        - name
        - brand
        - lat
        - lng
        - toiletType
        - requiresPermission
      properties:
        name:
          type: string
          description: コンビニ店舗名
          example: "セブン-イレブン 渋谷駅前店"
        brand:
          type: string
          enum:
            - seven_eleven
            - family_mart
            - lawson
            - mini_stop
            - daily_yamazaki
            - other
          description: コンビニブランド
        address:
          type: string
          description: 住所
        lat:
          type: number
          format: double
          description: 緯度
        lng:
          type: number
          format: double
          description: 経度
        imageKey:
          type: string
          description: S3にアップロード済みの画像キー
        toiletType:
          type: string
          enum: [shared, separated]
          description: "shared: 男女共用, separated: 男女別"
        requiresPermission:
          type: boolean
          description: 店員への使用許可が必要か
        note:
          type: string
          maxLength: 500
          description: その他自由記入欄

    Toilet:
      type: object
      required:
        - toiletId
        - name
        - brand
        - lat
        - lng
        - imageUrl
        - toiletType
        - requiresPermission
        - createdAt
      properties:
        toiletId:
          type: string
          format: uuid
        name:
          type: string
          description: コンビニ店舗名
          example: "セブン-イレブン 渋谷駅前店"
        brand:
          type: string
          enum:
            - seven_eleven
            - family_mart
            - lawson
            - mini_stop
            - daily_yamazaki
            - other
        address:
          type: string
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        imageUrl:
          type: string
          format: uri
          description: トイレ画像URL
        toiletType:
          type: string
          enum: [shared, separated]
          description: "shared: 男女共用, separated: 男女別"
        requiresPermission:
          type: boolean
        note:
          type: string
        voteCount:
          type: integer
          description: 投票数
        votes:
          type: array
          description: 投票一覧（getToilet で返す）
          items:
            $ref: "#/components/schemas/Vote"
        myVote:
          $ref: "#/components/schemas/Vote"
        createdBy:
          type: string
          description: 投稿者ユーザーID
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time